requirements.txt
pygame>=2.5.0
cairosvg>=2.7.0
Pillow>=10.0.0
